+++ legacy-data-access-services/src/test/java/gov/ca/cwds/cms/data/access/service/impl/dbDependentSuite/DBTestSuite.java
@@ -2,8 +2,8 @@ package gov.ca.cwds.cms.data.access.service.impl.dbDependentSuite;
 
 import gov.ca.cwds.cms.data.access.service.impl.dbDependent.CsecHistoryServiceTest;
 import gov.ca.cwds.cms.data.access.service.impl.dbDependent.SafetyAlertServiceTest;
-import gov.ca.cwds.cms.data.access.service.impl.relationships.dbDependent.R08840DBTest;
-import gov.ca.cwds.cms.data.access.service.impl.relationships.dbDependent.R08861DBTest;
+import gov.ca.cwds.cms.data.access.service.impl.clientrelationship.R08840DBTest;
+import gov.ca.cwds.cms.data.access.service.impl.clientrelationship.R08861DBTest;
 import gov.ca.cwds.cms.data.access.service.impl.tribalmembershipverification.dbDependent.R01128Test;
 import gov.ca.cwds.security.utils.PrincipalUtils;
 import org.junit.ClassRule;
